+++ fiber/src/fiber.mli
@@ -105,11 +105,21 @@ val parallel_iter_seq : 'a Seq.t -> f:('a -> unit t) -> unit t
 
 val sequential_iter_seq : 'a Seq.t -> f:('a -> unit t) -> unit t
 
-(** Provide efficient parallel iter/map functions for maps. *)
-module Make_map_traversals (Map : Map.S) : sig
-  val parallel_iter : 'a Map.t -> f:(Map.key -> 'a -> unit t) -> unit t
+(** Provide an efficient parallel map function for maps. *)
+module Make_parallel_map (S : sig
+  type 'a t
 
-  val parallel_map : 'a Map.t -> f:(Map.key -> 'a -> 'b t) -> 'b Map.t t
+  type key
+
+  val empty : _ t
+
+  val is_empty : _ t -> bool
+
+  val to_list : 'a t -> (key * 'a) list
+
+  val mapi : 'a t -> f:(key -> 'a -> 'b) -> 'b t
+end) : sig
+  val parallel_map : 'a S.t -> f:(S.key -> 'a -> 'b t) -> 'b S.t t
 end
 [@@inline always]